// src/types/api.ts
export interface FileEntry {
  name: string
  isDirectory: boolean
  path: string
  size: number
  mtime: string
}